+++ Test/EditorConfigTest/Helpers/Samples.cs
@@ -18,5 +18,10 @@ csharp_new_line_before_open_brace = accessors, indexers";
 dotnet_style_qualification_for_event = true";
 
         public const string Suppression = @"# Suppress: EC101 EC102";
+
+        public const string NamingRules = @"[*.cs]
+dotnet_naming_rule.foo.severity = warning
+dotnet_naming_rule.foo.symbols = fooSymbolsTitle
+dotnet_naming_rule.foo.style = fooStyleTitle";
     }
 }