+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Road map
 
+- [x] .NET naming rules basic Intellisense and validation
+- [ ] .NET naming rules dynamic validation
+- [ ] .NET naming rules snippet insertion
+- [ ] .NET naming rules dynamic Intellisense
 - [ ] Add item template for .editorconfig files
 
 Features that have a checkmark are complete and available for